+++ apps/desktop/native/crates/proxy/src/main.rs
@@ -1,15 +1,24 @@
-use anyhow::Result;
-use std::{
-    io::{self, stdin, Read, Write},
-    time::Duration,
-};
+use std::{io::stdin, time::Duration};
 
 use env_logger::Env;
 use log::debug;
 use tokio::time::sleep;
 
 mod ipc;
+mod native_messaging;
 
+/// Bitwarden IPC Proxy.
+///
+/// This proxy allows browser extensions to communicate with a desktop application using Native
+/// Messaging. This method allows an extension to send and receive messages through the use of
+/// stdin/stdout streams.
+///
+/// However, this also requires the browser to start the process in order for the communication to
+/// occur. To overcome this limitation, we implement Inter-Process Communication (IPC) to establish
+/// a stable communication channel between the proxy and the running desktop application.
+///
+/// Browser extension <-[native messaging]-> proxy <-[ipc]-> desktop
+///
 #[tokio::main]
 async fn main() {
     env_logger::Builder::from_env(Env::default().default_filter_or("debug")).init();
@@ -22,17 +31,20 @@ async fn main() {
     eprintln!();
     eprintln!("Starting Bitwarden IPC Proxy.");
 
+    // Setup two channels, one for sending messages to the desktop application and one for receiving messages
     let (in_tx, in_rx) = tokio::sync::mpsc::channel(32);
     let (out_tx, mut out_rx) = tokio::sync::mpsc::channel(32);
 
-    ipc::start(out_tx, in_rx);
+    tokio::spawn(async move {
+        ipc::start(out_tx, in_rx).await;
+    });
 
     // Receive messages from IPC and print to STDOUT.
     tokio::spawn(async move {
         loop {
             if let Some(msg) = out_rx.recv().await {
                 debug!("OUT: {}", msg);
-                write(msg);
+                native_messaging::write(msg);
             }
             sleep(Duration::from_millis(100)).await;
         }
@@ -40,10 +52,10 @@ async fn main() {
 
     // Listen to stdin and send messages to ipc processor.
     loop {
-        match read(stdin()) {
+        match native_messaging::read(stdin()) {
             Ok(msg) => {
                 if msg.len() == 0 {
-                    sleep(Duration::from_secs(5)).await;
+                    sleep(Duration::from_secs(1)).await;
                     continue;
                 }
 
@@ -60,29 +72,4 @@ async fn main() {
     }
 
     eprintln!("Exiting.");
-}
-
-/// Write a message to stdout. The message is prefixed with its length.
-///
-/// https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Native_messaging
-fn write(message: String) {
-    let bytes = message.as_bytes();
-    let header: [u8; 4] = (bytes.len() as u32).to_ne_bytes();
-    io::stdout().write(&header).ok();
-    io::stdout().write(bytes).ok();
-    io::stdout().flush().ok();
-}
-
-/// Read input from stdin. The input is expected to be prefixed with its length.
-///
-/// https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Native_messaging
-pub fn read<R: Read>(mut input: R) -> Result<Vec<u8>> {
-    let mut buf = [0; 4];
-    let length = input
-        .read_exact(&mut buf)
-        .map(|()| u32::from_ne_bytes(buf))?;
-
-    let mut buffer = vec![0; length as usize];
-    input.read_exact(&mut buffer)?;
-    Ok(buffer)
 }
